requirements.txt
pymysql
cryptography~=41.0.7
python-dotenv~=1.0.0
pip~=23.2.1
requests~=2.31.0
mysql~=0.0.3
certifi~=2023.11.17
mysqlclient~=2.2.1
idna~=3.6
urllib3~=2.1.0
pycparser~=2.21
mysql-connector-python